+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Changes in version 1.4**
 
+ - Added support for read/write options in the record-based layout, see RecordLayoutOptions class.
  - Fields that are completely blank in the data file won't be included in the resulting map even if their definition says they can't be trimmed.
  - Added support for NAACCR reserved fields; this feature is only available for NAACCR 16 and later.
  - This library now requires Java 8.